added App component's render test
