+++ src/index.js
@@ -1,32 +1,11 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import { BrowserRouter as Router } from "react-router-dom";
-import thunk from "redux-thunk";
-import logger from "redux-logger";
-import MuiThemeProvider from "material-ui/styles/MuiThemeProvider";
 import registerServiceWorker from "./registerServiceWorker";
 import "./index.css";
 import App from "./components/App";
-import { Provider } from "react-redux";
-import { createStore, applyMiddleware, compose  } from "redux";
-import appReducer from "./reducers/appReducer";
-
-
-const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
-const store = createStore(
-  appReducer,
-  {},
-  composeEnhancers(applyMiddleware(thunk, logger))
-);
 
 ReactDOM.render(
-  <Provider store={store}>
-    <Router>
-      <MuiThemeProvider>
-        <App />
-      </MuiThemeProvider>
-    </Router>
-  </Provider>,
+  <App />,
   document.getElementById("root")
 );
 registerServiceWorker();
